FastAPI/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
passlib
python-jose
pyjwt
python-multipart
requests